Moving test fixtures for reuse
