+++ ansible_base/lib/testing/fixtures.py
@@ -11,7 +11,7 @@ from cryptography.hazmat.primitives.asymmetric import rsa
 from cryptography.x509.oid import NameOID
 from rest_framework.test import APIClient
 
-from ansible_base.lib.testing.util import copy_fixture
+from ansible_base.lib.testing.util import copy_fixture, delete_authenticator
 
 
 @pytest.fixture
@@ -210,3 +210,36 @@ def rsa_keypair_with_cert(rsa_keypair_factory):
     certificate_bytes = certificate.public_bytes(serialization.Encoding.PEM).decode("utf-8")
     RSAKeyPairWithCert = namedtuple("RSAKeyPairWithCert", ["private", "public", "certificate"])
     return RSAKeyPairWithCert(private=rsa_keypair.private, public=rsa_keypair.public, certificate=certificate_bytes)
+
+
+@pytest.fixture
+def ldap_configuration():
+    return {
+        "SERVER_URI": ["ldap://ldap06.example.com:389"],
+        "BIND_DN": "cn=ldapadmin,dc=example,dc=org",
+        "BIND_PASSWORD": "securepassword",
+        "START_TLS": False,
+        "CONNECTION_OPTIONS": {"OPT_REFERRALS": 0, "OPT_NETWORK_TIMEOUT": 30},
+        "USER_SEARCH": ["ou=users,dc=example,dc=org", "SCOPE_SUBTREE", "(cn=%(user)s)"],
+        "USER_DN_TEMPLATE": "cn=%(user)s,ou=users,dc=example,dc=org",
+        "USER_ATTR_MAP": {"email": "mail", "last_name": "sn", "first_name": "givenName"},
+        "GROUP_SEARCH": ["ou=groups,dc=example,dc=org", "SCOPE_SUBTREE", "(objectClass=groupOfNames)"],
+        "GROUP_TYPE": "MemberDNGroupType",
+        "GROUP_TYPE_PARAMS": {"name_attr": "cn", "member_attr": "member"},
+    }
+
+
+@pytest.fixture
+def ldap_authenticator(ldap_configuration):
+    from ansible_base.authentication.models import Authenticator
+
+    authenticator = Authenticator.objects.create(
+        name="Test LDAP Authenticator",
+        enabled=True,
+        create_objects=True,
+        remove_users=True,
+        type="ansible_base.authentication.authenticator_plugins.ldap",
+        configuration=ldap_configuration,
+    )
+    yield authenticator
+    delete_authenticator(authenticator)
