+++ test_app/tests/conftest.py
@@ -19,7 +19,7 @@ from rest_framework.request import Request
 from rest_framework.test import force_authenticate
 
 from ansible_base.lib.testing.fixtures import *  # noqa: F403, F401
-from ansible_base.lib.testing.util import copy_fixture
+from ansible_base.lib.testing.util import copy_fixture, delete_authenticator
 from ansible_base.rbac import permission_registry
 from ansible_base.rbac.models import RoleDefinition
 from test_app import models
@@ -63,23 +63,6 @@ def test_migrations_okay(*args, **kwargs):
 
 
 post_migrate.connect(test_migrations_okay)
-
-
-def delete_authenticator(authenticator):
-    from django.conf import settings
-
-    from ansible_base.authentication.models import AuthenticatorUser
-
-    for au in AuthenticatorUser.objects.filter(provider=authenticator):
-        try:
-            # The tests are very sensitive to the SYSTEM_USER being removed so we won't delete that user
-            if au.username != settings.SYSTEM_USERNAME:
-                au.user.delete()
-        except Exception:
-            # Its possible that something else already delete the user if a user was multi linked somehow
-            pass
-        au.delete()
-    authenticator.delete()
 
 
 @pytest.fixture
@@ -312,39 +295,6 @@ def oidc_authenticator(oidc_configuration):
         remove_users=True,
         type="ansible_base.authentication.authenticator_plugins.oidc",
         configuration=oidc_configuration,
-    )
-    yield authenticator
-    delete_authenticator(authenticator)
-
-
-@pytest.fixture
-def ldap_configuration():
-    return {
-        "SERVER_URI": ["ldap://ldap06.example.com:389"],
-        "BIND_DN": "cn=ldapadmin,dc=example,dc=org",
-        "BIND_PASSWORD": "securepassword",
-        "START_TLS": False,
-        "CONNECTION_OPTIONS": {"OPT_REFERRALS": 0, "OPT_NETWORK_TIMEOUT": 30},
-        "USER_SEARCH": ["ou=users,dc=example,dc=org", "SCOPE_SUBTREE", "(cn=%(user)s)"],
-        "USER_DN_TEMPLATE": "cn=%(user)s,ou=users,dc=example,dc=org",
-        "USER_ATTR_MAP": {"email": "mail", "last_name": "sn", "first_name": "givenName"},
-        "GROUP_SEARCH": ["ou=groups,dc=example,dc=org", "SCOPE_SUBTREE", "(objectClass=groupOfNames)"],
-        "GROUP_TYPE": "MemberDNGroupType",
-        "GROUP_TYPE_PARAMS": {"name_attr": "cn", "member_attr": "member"},
-    }
-
-
-@pytest.fixture
-def ldap_authenticator(ldap_configuration):
-    from ansible_base.authentication.models import Authenticator
-
-    authenticator = Authenticator.objects.create(
-        name="Test LDAP Authenticator",
-        enabled=True,
-        create_objects=True,
-        remove_users=True,
-        type="ansible_base.authentication.authenticator_plugins.ldap",
-        configuration=ldap_configuration,
     )
     yield authenticator
     delete_authenticator(authenticator)
